took imageDrop  handler to shared service
